Controllers and functionality
Now the controllers exist and works well. The methods which was implementes are the methods are used but probably there is necessary to implement more.
Right now the project is working and if you save and user, the postman returns the recommended credit cards depends on how you save the data in the data base.