Modified approach
The project has changed its approach.
Now there are only 2 entities, Usuario has been disabled and probably eliminated if I cant find a certain function.
All the controller have now encapsulation.
The recommendation of credit cards is now part of the controller of Perfil and it works with params instead of variables.